Documentation of "ACS" library for R
Library includes acs.fetch() which will download American Communities Survey data from the U.S. Census.  You have to register with the Census Bureau and get an API key.  This wasn't as helpful as we thought, because the API key doesn't work for summary tables (like S1501, S1701, S1903, and S2301), and we get a lot of data from summary tables.  The data is available from other tables, but takes a lot of work to summarize.